Add /system/info endpoint for faas-nomad
The purpose of this change is to make faas-nomad compatible with the
latest version of the API Gateway in OpenFaaS. Each provider should
be able to advertise the underlying provider and the version of the
code. A new version of faas-provider (0.6) is vendored into the repo
which brings in an optional health endpoint (turned off by default).
The version is being passed to the provider from main.go.

A subsequent change should implement the health endpoint.

A unit test was added for the info endpoint using the format found
for the other endpoints.

The advertised information in /system/info will be consumed by the
API gateway and logged for help debugging issues. It can also be
used to turn on/off UI features such as the "network" field which
does not make sense for all providers.

Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>